pyproject.toml
[project]
name = "frp-manager"
version = "0.1.0"
description = "FRP 内网穿透管理工具，支持服务端/客户端"
requires-python = ">=3.10"
dependencies = [
    "click>=8.3.1",
    "pyyaml>=6.0.3",
]

[project.scripts]
frp-ctl = "frp_manager.cli:main"

[tool.uv]
package = true

[tool.uv.sources]
